requirements.txt
pyrogram==2.0.106
py-tgcalls==1.0.2
TgCrypto==1.2.5
youtube-search-python==1.6.6
youtube-dl==2021.12.17
yt-dlp==2023.4.13
youtube-search==2.1.2
python-dotenv==1.0.0
motor==3.2.0
psutil==5.9.6
aiohttp==3.8.5
aiofiles==23.2.1
pillow==10.0.1
ffmpeg-python==0.2.0
dnspython==2.4.2
heroku3==5.2.1
requests==2.31.0
numpy==1.24.3
spotipy==2.23.0
